fix problem asking for password multiple times refactoring
