+++ .gitignore
@@ -15,3 +15,5 @@
 # vendor/
 /bin/
 /.idea/
+/INTERN.txt
+/version_git_tag
