+++ Makefile
@@ -1,4 +1,7 @@
-build:
+create_git_tag:
+	git describe --abbrev=0 --tags > version_git_tag
+
+build: create_git_tag
 	mkdir -p bin
 	go build -ldflags="-s -w" -o bin/ssh-key-exchange .
 
@@ -14,7 +17,7 @@ build_windows_amd64:
 	mkdir -p bin/windows_amd64
 	GOOS=windows GOARCH=amd64 go build -ldflags="-s -w" -o bin/windows_amd64/ssh-key-exchange.exe .
 
-build-all: build_darwin_amd64 build_linux_amd64 build_windows_amd64
+build-all: create_git_tag build_darwin_amd64 build_linux_amd64 build_windows_amd64
 
 install: build
 	go install
